Project SoFiA
Java application designed to visualize the look of the Fraunhofer diffraction patterns in the far-field, created on different (circular and rectangular) apertures. Users can vary aperture dimensions along with the incident light wavelength and the distance from the image observation. There is also a possibility of animating shifts of aperture parameters in real-time.